initial settings refactor #1
