+++ src/formApps/_AutorecMenu/components/index.js
@@ -1,1 +1,2 @@
+export { default as AAGameSettings }  from "./settings/AAGameSettings.svelte";
 export { default as AutorecAppShell } from "./AutorecAppShell.svelte";
